Fix shellcheck issue

+++ buildpacks/npm/lib/build.sh
@@ -182,7 +182,7 @@ install_or_reuse_node_modules() {
 		info "Reusing node modules"
 		cp -r "$layer_dir" "$build_dir/node_modules"
 	else
-    rm -rf "$layer_dir"/*
+		rm -rf "${layer_dir:?}"/*
 		echo "cache = true" >"${layer_dir}.toml"
 
 		{
